rename aux files
