+++ src/components/Navigation/Search/Search.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Aux from "../../../hoc/Aux/Aux";
+import Aux from "../../../hoc/_Aux/_Aux";
 
 const search = (props) => {
   return (
