+++ src/containers/Home/Home.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import Aux from "../../hoc/Aux/Aux";
+import Aux from "../../hoc/_Aux/_Aux";
